Use HKDF, ChaCha20, and backends
In changing things over to use HKDF and ChaCha20 it became obvious that a separate backend implementation would be necessary. I moved the actual storage of blobs into separate interfaces as defined by Backend. This will allow for unit testing the blobstore without accessing S3. The S3 code will still need to be tested, but only as the S3 backend implementation changes. In these cases an S3 double can be used or an actual S3 bucket.